+++ ft_convp.c
@@ -6,7 +6,7 @@
 /*   By: mde-laga <mde-laga@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/02/07 15:59:30 by mde-laga          #+#    #+#             */
-/*   Updated: 2019/02/09 16:01:50 by mde-laga         ###   ########.fr       */
+/*   Updated: 2019/02/10 15:48:25 by mde-laga         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ static char	*ft_getstring(t_prin *prin)
 
 	ptr = va_arg(prin->ap, void*);
 	str = ft_strdup("0x");
-	str = ft_strjfree(str, ft_llutoa_base((unsigned long long)ptr, 16, 'a'));
+	str = ft_strjfree(str, ft_llutoa_base((uint64_t)ptr, 16, 'a'));
 	return (str);
 }
 
@@ -44,7 +44,7 @@ void		ft_convp(t_prin *prin)
 	char	*ret;
 
 	ret = ft_getstring(prin);
-	if ((unsigned long)prin->field > ft_strlen(ret))
+	if ((uint64_t)prin->field > ft_strlen(ret))
 		ret = ft_add(prin, ret);
 	prin->ret += ft_strlen(ret);
 	prin->output = ft_strjfree(prin->output, ret);
