+++ ft_lltoa.c
@@ -6,13 +6,13 @@
 /*   By: mde-laga <mde-laga@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/01/30 15:02:34 by mde-laga          #+#    #+#             */
-/*   Updated: 2019/02/09 16:02:56 by mde-laga         ###   ########.fr       */
+/*   Updated: 2019/02/10 15:50:10 by mde-laga         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int		ft_len(long long nb)
+int		ft_len(int64_t nb)
 {
 	int len;
 
@@ -24,7 +24,7 @@ int		ft_len(long long nb)
 	return (len + 1);
 }
 
-void	ft_recur(char *str, long long nb, int i)
+void	ft_recur(char *str, int64_t nb, int i)
 {
 	int neg;
 
@@ -34,7 +34,7 @@ void	ft_recur(char *str, long long nb, int i)
 	str[i] = nb % 10 * neg + '0';
 }
 
-char	*ft_lltoa(long long nb)
+char	*ft_lltoa(int64_t nb)
 {
 	char	*str;
 	int		i;
